elt-pipeline/requirements.txt
apache-airflow-providers-google
pandas
supabase
google-cloud-bigquery
python-dotenv